chore(TCOMP-680): rename design.manager package to design.extension
