+++ component-runtime-design-extension/src/main/java/org/talend/sdk/component/design/extension/DesignContainerListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.talend.sdk.component.design.manager;
+package org.talend.sdk.component.design.extension;
 
 import java.util.stream.Stream;
 
@@ -40,11 +40,12 @@ public class DesignContainerListener implements ContainerListenerExtension {
         componentRegistry.getComponents().values().stream()
                 .flatMap(family -> Stream.concat( //
                         family.getPartitionMappers().values().stream(), //
-                        family.getProcessors().values().stream()) //
-                ).forEach(meta -> meta.set(DesignModel.class, new DesignModel( //
-                        meta.getId(), //
-                        meta.getInputFlows(), //
-                        meta.getOutputFlows()))); //
+                        family.getProcessors().values().stream())) //
+                .forEach(meta -> meta.set(DesignModel.class,
+                        new DesignModel( //
+                                meta.getId(), //
+                                meta.getInputFlows(), //
+                                meta.getOutputFlows()))); //
     }
 
     /**
