+++ component-runtime-design-extension/src/main/java/org/talend/sdk/component/design/extension/DesignModel.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.talend.sdk.component.design.manager;
+package org.talend.sdk.component.design.extension;
 
 import java.util.Collection;
