+++ component-runtime-design-extension/src/main/resources/META-INF/services/org.talend.sdk.component.runtime.manager.spi.ContainerListenerExtension
@@ -1,1 +1,1 @@
-org.talend.sdk.component.design.manager.DesignContainerListener
+org.talend.sdk.component.design.extension.DesignContainerListener
